Moved pipes to pipe folder, updated settings-label to match constructor
