+++ src/app/shared/pipes/settings-label.pipe.spec.ts
@@ -1,0 +1,10 @@
+import { SettingsLabelPipe } from './settings-label.pipe';
+import { ConvertUnitsService } from '../convert-units/convert-units.service';
+
+describe('SettingsLabelPipe', () => {
+  it('create an instance', () => {
+    const convertUnitsService = new ConvertUnitsService();
+    const pipe = new SettingsLabelPipe(convertUnitsService);
+    expect(pipe).toBeTruthy();
+  });
+});
