--- src/app/shared/settings-label.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { SettingsLabelPipe } from './settings-label.pipe';
-
-describe('SettingsLabelPipe', () => {
-  it('create an instance', () => {
-    const pipe = new SettingsLabelPipe();
-    expect(pipe).toBeTruthy();
-  });
-});
